# backend/hollow/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/user/info/{id}:
        get:
            tags:
                - Users
            description: 获取用户资料
            operationId: Users_GetUser
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.GetUserReply'
    /api/user/login:
        post:
            tags:
                - Users
            description: 登录
            operationId: Users_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.LoginUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.LoginUserReply'
    /api/user/mfa:
        get:
            tags:
                - Users
            description: MFA获取秘钥二维码
            operationId: Users_MFAGetQRCode
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.MFAGetQRCodeReply'
        post:
            tags:
                - Users
            description: MFA激活
            operationId: Users_MFAActivate
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.MFAActivateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.MFAActivateReply'
        delete:
            tags:
                - Users
            description: MFA解绑
            operationId: Users_MFACancel
            parameters:
                - name: code
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.MFACancelReply'
    /api/user/rebind/phone:
        post:
            tags:
                - Users
            operationId: Users_ReBindPhone
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.ReBindPhoneRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.ReBindPhoneReply'
    /api/user/register:
        post:
            tags:
                - Users
            description: 注册
            operationId: Users_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.RegisterUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.RegisterUserReply'
    /api/user/register/shortmsg:
        post:
            tags:
                - Users
            description: 发送手机验证码
            operationId: Users_SendShortMsg
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.SendShortMsgRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.SendShortMsgReply'
components:
    schemas:
...
        user.v1.GetUserReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int64
                msg:
                    type: string
                data:
                    $ref: '#/components/schemas/user.v1.User'
        user.v1.LoginUserReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int64
                msg:
                    type: string
                token:
                    type: string
                data:
                    $ref: '#/components/schemas/user.v1.User'
        user.v1.LoginUserRequest:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
                code:
                    type: string
        user.v1.MFA:
            type: object
            properties:
                qrlink:
                    type: string
                secret:
                    type: string
        user.v1.MFAActivateReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int64
                msg:
                    type: string
        user.v1.MFAActivateRequest:
            type: object
            properties:
                code:
                    type: string
                secret:
                    type: string
        user.v1.MFACancelReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int64
                msg:
                    type: string
        user.v1.MFAGetQRCodeReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int64
                msg:
                    type: string
                data:
                    $ref: '#/components/schemas/user.v1.MFA'
...
        user.v1.ReBindPhoneReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int64
                msg:
                    type: string
        user.v1.ReBindPhoneRequest:
            type: object
            properties:
                phone:
                    type: string
                code:
                    type: string
                mfacode:
                    type: string
        user.v1.RegisterUserReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int64
                msg:
                    type: string
                data:
                    $ref: '#/components/schemas/user.v1.User'
        user.v1.RegisterUserRequest:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
                email:
                    type: string
                phone:
                    type: string
                code:
                    type: string
        user.v1.SendShortMsgReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int64
                msg:
                    type: string
                data:
                    $ref: '#/components/schemas/user.v1.ShortMsg'
        user.v1.SendShortMsgRequest:
            type: object
            properties:
                phone:
                    type: string
        user.v1.ShortMsg:
            type: object
            properties:
                code:
                    type: string
                message:
                    type: string
                bizid:
                    type: string
                requestid:
                    type: string
        user.v1.User:
            type: object
            properties:
                userid:
                    type: integer
                    format: int64
                username:
                    type: string
                status:
                    type: integer
                    format: int64
                email:
                    type: string
                nickname:
                    type: string
                createdAt:
                    type: integer
                    format: int64
                phone:
                    type: integer
                    format: int64
                mfaEnabled:
                    type: boolean
...